refactor: data fetching in flashcards page
* previously violated the NextJS convention of using server actions only for mutating data. Now the known questions and question id's arrays are being fetched directly in the page component
* fixed privacy policy link in the footer
