+++ src/app/(games)/flashcards/[code]/page.tsx
@@ -1,10 +1,38 @@
 import Flashcards from '@/components/flashcards/Flashcards'
 import PageTitle from '@/components/ui/PageTitle'
-import getKnownQuestions from '@/lib/supabase/get-known-questions'
-import getQuestionIdsList from '@/lib/supabase/get-questions-id-list'
 import getUser from '@/lib/supabase/get-user'
 import { createClient } from '@/lib/supabase/server'
 import { notFound, redirect } from 'next/navigation'
+
+async function getKnownQuestions(userId: string, examId: number) {
+  const supabase = createClient()
+  const { data, error } = await supabase
+    .from('flashcards')
+    .select('question_id_array')
+    .eq('user_id', userId)
+    .eq('exam_id', examId)
+    .single()
+  if (!data || error) {
+    return []
+  } else {
+    return data.question_id_array
+  }
+}
+
+async function getQuestionIdArray(examId: number) {
+  const supabase = createClient()
+
+  const { data, error } = await supabase
+    .from('questions')
+    .select('id')
+    .eq('exam_id', examId)
+
+  if (!data || data.length < 1 || error) {
+    throw new Error('Failed to fetch questions from the database')
+  } else {
+    return data.map((obj) => obj.id)
+  }
+}
 
 export default async function FlashcardsPage({ params }: { params: { code: string } }) {
   const supabase = createClient()
@@ -26,7 +54,7 @@ export default async function FlashcardsPage({ params }: { params: { code: strin
   }
 
   const knownQuestions = await getKnownQuestions(user.id, data.id)
-  const questionIds = await getQuestionIdsList(data.id)
+  const questionIds = await getQuestionIdArray(data.id)
 
   return (
     <>
